refactor: integrate google platform login handler
