--- api/auth/gmail-login-url-handler/Dockerfile
@@ -1,9 +1,0 @@
-FROM public.ecr.aws/lambda/python:3.10
-
-COPY requirements.txt .
-
-RUN pip install -r requirements.txt
-
-COPY . .
-
-CMD [ "lambda_function.handler" ]
