--- api/auth/google-drive-login-url-handler/requirements.txt
@@ -1,3 +1,0 @@
-google-auth
-google-auth-oauthlib
-boto3
